feat: Remove the need for the `downcast` function.
The `downcast` function is removed in this patch because it was
incompatible with JS code minifier.

Instead, we now rely on the new `wasm_bindgen::convert::TryFromJsValue`
API.

This patch also replaces `&Array` by `Vec<T>`. It's fine, except that JS
arrays that are sent to Rust are going to be dropped after the functions
return. It's quite unsual for JS, but documentations have been updated
accordingly.