Fix a doc comment incorrectly stating that the function returns a promise
In cbd7ca9d4 the Device.verificationRequest method was converted from an
async method to a sync one, this was done because the underlying SDK
method has received the same treatment.

The doc comment was not updated to reflect this change and claimed that
it's still returning a promise. This patch fixes this mistake.